Make compatible with Gradle 4.3 (#2)
This was done by using Groovy's `InvokerHelper` to wrap the check for
`task.archiveFile`. In Gradle 5.0 and older, `AbstractArchiveTask`
exposed the output with `task.archivePath`. If the former can't be
found, it will try to find the latter. If neither can be found, both
exceptions will be thrown with the latter being included as a suppressed
exception to the former.

This plugin still has a lot of problems with configuration cache, so a
2.0 will be created in due time that will target 9.0.0 (or 7.3) to
address those problems. There are a lot of remedies for configuration
cache that only exist in newer versions of Gradle.